requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
openai
python-dotenv
requests
Pillow
numpy
opencv-python
torch
torchvision
imagehash
